pyproject.toml
[tool.poetry]
name = "mxcubeweb"
version = "1.113.0"
license = "LGPL-3.0-or-later"
description = "MXCuBE Web user interface"
authors = ["The MXCuBE collaboration <mxcube@esrf.fr>"]
maintainers = [
    "MXCuBE collaboration <mxcube@esrf.fr>",
]
readme = "README.md"
homepage = "http://github.com/mxcube/mxcubeweb"
repository = "http://github.com/mxcube/mxcubeweb"
documentation = ""
keywords = ["mxcube", "mxcube3", "mxcubeweb"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Topic :: Scientific/Engineering",
]
packages = [{include = "mxcubeweb", from = "."}]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
Flask = "^2.2.2"
werkzeug = "2.3.7"
Flask-SocketIO = "^5.3.2"
flask-sqlalchemy = "^3.0.2"
Flask-Security-Too = "^5.0.2"
gevent-websocket = "0.10.1"
spectree = "1.2.1"
mock = "^4.0.3"
"ruamel.yaml" = "^0.17.21"
scandir = "^1.10.0"
redis = "^4.3.5"
devtools = "^0.10.0"
jsonschema = "^4.17.1"
pydantic = "^1.10.2"
PyDispatcher = "^2.0.6"
pytz = "^2022.6"
tzlocal = "^4.2"
mxcubecore = "1.34.0"
mxcube_video_streamer = "1.0.0"
bcrypt = "^4.0.1"

[tool.poetry.dev-dependencies]
pre-commit = "2.20.0"
pytest = "7.1.3"
pytest-cov = "4.0.0"

[tool.poetry.group.docs.dependencies]
myst-parser = "^2.0.0"
sphinx = "<7.2"
sphinxcontrib-httpdomain = "^1.8.1"
furo = "^2023.9.10"

[tool.black]
line-length = 88
preview = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | test
  | ui
  | docs
)/
'''

[build-system]
requires = ["poetry-core>=1.0.0", "setuptools>=45", "setuptools_scm[toml]>=6.2"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mxcubeweb-server = 'mxcubeweb:main'
